remove srcset when replacing
On modern websites, images often use the srcset attribute. If they are loaded using this, it may not work to just change the src attribute, because the browser loaded one of the srcset images. By removing the srcset on replace, the browser is forced to take the image from src. Now it works again.